frontend: extract shared coin helpers
Move coin classification and coin-display helpers out of the account route
utilities into a shared coin utility module. This keeps account route helpers
focused on account lookup, grouping, route checks, and script names while
letting account views, market flows, send/receive UI, and the ETH activity
hook share the same coin-code rules.
