+++ frontends/web/src/components/groupedaccountselector/groupedaccountselector.tsx
@@ -9,7 +9,8 @@ import { USBSuccess, ChevronDownDark } from '@/components/icon';
 import { Badge } from '@/components/badge/badge';
 import { InsuredShield } from '@/routes/account/components/insuredtag';
 import { RatesContext } from '@/contexts/RatesContext';
-import { getAccountsByKeystore, getDisplayedCoinUnit } from '@/routes/account/utils';
+import { getAccountsByKeystore } from '@/routes/account/utils';
+import { getDisplayedCoinUnit } from '@/utils/coin';
 import { Dropdown, TOption as TDropdownOption, TGroupedOption as TDropdownGroupedOption } from '@/components/dropdown/dropdown';
 import { createGroupedOptions, getBalancesForGroupedAccountSelector } from './services';
 import { AmountWithUnit } from '../amount/amount-with-unit';
