+++ frontends/web/src/routes/account/addresses/dialog/skipped-dialog.tsx
@@ -2,7 +2,7 @@
 
 import { useTranslation } from 'react-i18next';
 import { Dialog } from '@/components/dialog/dialog';
-import { getAddressURIPrefix } from '@/routes/account/utils';
+import { getAddressURIPrefix } from '@/utils/coin';
 import { VerifyAddressDialogContent } from '../../components/verify-address-dialog-content';
 import { AddressNotFoundDialog } from './address-not-found-dialog';
 import { TDialogProps } from './types';
