+++ frontends/web/src/routes/account/addresses/dialog/verify-on-device-dialog.tsx
@@ -5,7 +5,7 @@ import { Message } from '@/components/message/message';
 import { Dialog } from '@/components/dialog/dialog';
 import { Button } from '@/components/forms';
 import { UseBackButton } from '@/hooks/backbutton';
-import { getAddressURIPrefix } from '@/routes/account/utils';
+import { getAddressURIPrefix } from '@/utils/coin';
 import { VerifyAddressDialogContent } from '../../components/verify-address-dialog-content';
 import { AddressNotFoundDialog } from './address-not-found-dialog';
 import { TDialogProps } from './types';
