+++ frontends/web/src/routes/account/info/info.tsx
@@ -6,7 +6,8 @@ import { useTranslation } from 'react-i18next';
 import { useSync } from '@/hooks/api';
 import { useMountedRef } from '@/hooks/mount';
 import { TAccount, AccountCode, TStatus, getStatus, exportAccount, getTransactionList, TTransactions } from '@/api/account';
-import { findAccount, isBitcoinBased, isMessageSigningSupported } from '@/routes/account/utils';
+import { findAccount } from '@/routes/account/utils';
+import { isBitcoinBased, isMessageSigningSupported } from '@/utils/coin';
 import { TDevices } from '@/api/devices';
 import { Header, Main } from '@/components/layout';
 import { View, ViewContent } from '@/components/view/view';
