+++ frontends/web/src/routes/account/info/xpub-detail.tsx
@@ -6,9 +6,9 @@ import { useTranslation } from 'react-i18next';
 import { useLoad } from '@/hooks/api';
 import { getInfo, TAccount, AccountCode, ScriptType, TSigningConfiguration } from '@/api/account';
 import { findAccount } from '@/routes/account/utils';
+import { isBitcoinBased } from '@/utils/coin';
 import { GuidedContent, GuideWrapper, Header, Main } from '@/components/layout';
 import { View, ViewContent } from '@/components/view/view';
-import { isBitcoinBased } from '@/routes/account/utils';
 import { BitcoinBasedAccountInfoGuide } from './guide';
 import { MobileHeader } from '@/routes/settings/components/mobile-header';
 import { BackButton } from '@/components/backbutton/backbutton';
