+++ frontends/web/src/routes/account/receive/receive.tsx
@@ -9,7 +9,8 @@ import { UseBackButton } from '@/hooks/backbutton';
 import * as accountApi from '@/api/account';
 import { setAccountReceiveScriptType } from '@/api/backend';
 import { alertUser } from '@/components/alert/Alert';
-import { getScriptName, isEthereumBased } from '@/routes/account/utils';
+import { getScriptName } from '@/routes/account/utils';
+import { isEthereumBased } from '@/utils/coin';
 import { CopyableInput } from '@/components/copy/Copy';
 import { Dialog, DialogButtons, DialogScrollContent } from '@/components/dialog/dialog';
 import { Button, Radio } from '@/components/forms';
