+++ frontends/web/src/routes/account/utils.test.ts
@@ -4,12 +4,6 @@ import { describe, it, expect } from 'vitest';
 import { CoinCode, CoinUnit, TAccount } from '@/api/account';
 import {
   getAccountsByKeystore,
-  getAddressURIPrefix,
-  getCoinCode,
-  isBitcoinBased,
-  isBitcoinCoin,
-  isBitcoinOnly,
-  isMessageSigningSupported,
 } from './utils';
 
 
@@ -127,42 +121,5 @@ describe('utils/getAccountsByKeystore', () => {
     expect(result[0].accounts[2].code).toBe(accounts[6].code);
     // @ts-ignore noUncheckedIndexedAccess
     expect(result[1].accounts[3].code).toBe(accounts[5].code);
-  });
-});
-
-describe('utils/bitcoin coin helpers', () => {
-  it('treats rbtc coin codes as bitcoin-only and bitcoin-based', () => {
-    expect(isBitcoinOnly('rbtc')).toBe(true);
-    expect(isBitcoinBased('rbtc')).toBe(true);
-  });
-
-  it('treats rbtc unit as bitcoin coin', () => {
-    expect(isBitcoinCoin('RBTC' as CoinUnit)).toBe(true);
-  });
-
-  it('uses bitcoin URI prefix for rbtc', () => {
-    expect(getAddressURIPrefix('rbtc')).toBe('bitcoin:');
-  });
-
-  it('maps rbtc to canonical btc coin code', () => {
-    expect(getCoinCode('rbtc' as CoinCode)).toBe('btc');
-  });
-});
-
-describe('utils/isMessageSigningSupported', () => {
-  it('supports btc, tbtc and rbtc message signing', () => {
-    expect(isMessageSigningSupported('btc')).toBe(true);
-    expect(isMessageSigningSupported('tbtc')).toBe(true);
-    expect(isMessageSigningSupported('rbtc')).toBe(true);
-  });
-
-  it('supports sepolia and eth message signing', () => {
-    expect(isMessageSigningSupported('eth')).toBe(true);
-    expect(isMessageSigningSupported('sepeth')).toBe(true);
-  });
-
-  it('does not support litecoin message signing', () => {
-    expect(isMessageSigningSupported('ltc')).toBe(false);
-    expect(isMessageSigningSupported('tltc')).toBe(false);
   });
 });
