+++ frontends/web/src/routes/account/utils.ts
@@ -1,125 +1,12 @@
 // SPDX-License-Identifier: Apache-2.0
 
-import type { AccountCode, CoinCode, ScriptType, TAccount, TAccountBase, CoinUnit, TKeystore } from '@/api/account';
-import type { BtcUnit } from '@/api/coins';
+import type { AccountCode, CoinCode, ScriptType, TAccount, TAccountBase, TKeystore } from '@/api/account';
 
 export const findAccount = <T extends { code: AccountCode }>(
   accounts: T[],
   accountCode: AccountCode
 ): T | undefined => {
   return accounts.find(({ code }) => accountCode === code);
-};
-
-export const isBitcoinOnly = (coinCode: CoinCode): boolean => {
-  switch (coinCode) {
-  case 'btc':
-  case 'tbtc':
-  case 'rbtc':
-    return true;
-  default:
-    return false;
-  }
-};
-
-export const isBitcoinCoin = (coin: CoinUnit | undefined) => {
-  switch (coin) {
-  case 'BTC':
-  case 'TBTC':
-  case 'RBTC':
-  case 'sat':
-  case 'tsat':
-    return true;
-  default:
-    return false;
-  }
-};
-
-export const getDisplayedCoinUnit = (
-  coinCode: CoinCode,
-  coinUnit: CoinUnit,
-  btcUnit: BtcUnit | undefined,
-): CoinUnit => {
-  if (!isBitcoinOnly(coinCode) || btcUnit !== 'sat') {
-    return coinUnit;
-  }
-  return coinCode === 'tbtc' ? 'tsat' : 'sat';
-};
-
-export const isBitcoinBased = (coinCode: CoinCode): boolean => {
-  switch (coinCode) {
-  case 'btc':
-  case 'tbtc':
-  case 'rbtc':
-  case 'ltc':
-  case 'tltc':
-    return true;
-  default:
-    return false;
-  }
-};
-
-export const isBitcoinCoinBased = (coinCode: CoinUnit): boolean => {
-  switch (coinCode) {
-  case 'BTC':
-  case 'TBTC':
-  case 'RBTC':
-  case 'sat':
-  case 'tsat':
-  case 'LTC':
-  case 'TLTC':
-    return true;
-  default:
-    return false;
-  }
-};
-
-export const isEthereumBased = (coinCode: CoinCode): boolean => {
-  return coinCode === 'eth' || coinCode === 'sepeth' || coinCode.startsWith('eth-erc20-');
-};
-
-export const isMessageSigningSupported = (coinCode: CoinCode): boolean => {
-  switch (coinCode) {
-  case 'btc':
-  case 'tbtc':
-  case 'eth':
-  case 'sepeth':
-  case 'rbtc':
-    return true;
-  default:
-    return false;
-  }
-};
-
-export const getAddressURIPrefix = (coinCode?: CoinCode): string => {
-  switch (coinCode) {
-  case 'btc':
-  case 'tbtc':
-  case 'rbtc':
-    return 'bitcoin:';
-  case 'ltc':
-  case 'tltc':
-    return 'litecoin:';
-  default:
-    return '';
-  }
-};
-
-export const getCoinCode = (coinCode: CoinCode): CoinCode | undefined => {
-  switch (coinCode) {
-  case 'btc':
-  case 'tbtc':
-  case 'rbtc':
-    return 'btc';
-  case 'ltc':
-  case 'tltc':
-    return 'ltc';
-  case 'eth':
-  case 'sepeth':
-    return 'eth';
-  }
-  if (coinCode.startsWith('eth-erc20-')) {
-    return 'eth';
-  }
 };
 
 export const getScriptName = (scriptType: ScriptType): string => {
@@ -133,16 +20,6 @@ export const getScriptName = (scriptType: ScriptType): string => {
   case 'p2tr':
     return 'Taproot (bech32m, P2TR)';
   }
-};
-
-export const customFeeUnit = (coinCode: CoinCode): string => {
-  if (isBitcoinBased(coinCode)) {
-    return 'sat/vB';
-  }
-  if (isEthereumBased(coinCode)) {
-    return 'Gwei';
-  }
-  return '';
 };
 
 export type TAccountsByKeystore<T extends { keystore: TKeystore } = TAccount> = {
