+++ frontends/web/src/routes/market/bitrefill.tsx
@@ -7,7 +7,8 @@ import { MobileHeader } from '../settings/components/mobile-header';
 import { Spinner } from '@/components/spinner/Spinner';
 import { MarketGuide } from './guide';
 import { AccountCode, TAccount, proposeTx, sendTx, TTxInput, TTxProposalResult } from '@/api/account';
-import { findAccount, isBitcoinOnly } from '@/routes/account/utils';
+import { findAccount } from '@/routes/account/utils';
+import { isBitcoinOnly } from '@/utils/coin';
 import { useDarkmode } from '@/hooks/darkmode';
 import { useConfig } from '@/contexts/ConfigProvider';
 import { i18n } from '@/i18n/i18n';
