+++ frontends/web/src/routes/market/btcdirect.tsx
@@ -15,7 +15,8 @@ import { getURLOrigin } from '@/utils/url';
 import { Header } from '@/components/layout';
 import { MobileHeader } from '../settings/components/mobile-header';
 import { Spinner } from '@/components/spinner/Spinner';
-import { findAccount, isBitcoinOnly } from '@/routes/account/utils';
+import { findAccount } from '@/routes/account/utils';
+import { isBitcoinOnly } from '@/utils/coin';
 import { BTCDirectTerms } from '@/components/terms/btcdirect-terms';
 import { MarketGuide } from './guide';
 import { alertUser } from '@/components/alert/Alert';
