+++ frontends/web/src/routes/market/moonpay.tsx
@@ -12,7 +12,8 @@ import { Header } from '@/components/layout';
 import { MobileHeader } from '../settings/components/mobile-header';
 import { Message } from '@/components/message/message';
 import { Spinner } from '@/components/spinner/Spinner';
-import { findAccount, isBitcoinOnly } from '@/routes/account/utils';
+import { findAccount } from '@/routes/account/utils';
+import { isBitcoinOnly } from '@/utils/coin';
 import { MoonpayTerms } from '@/components/terms/moonpay-terms';
 import { useVendorIframeResizeHeight, useVendorTerms } from '@/hooks/vendor-iframe';
 import style from './iframe.module.css';
