+++ frontends/web/src/routes/market/swap/components/input-with-account-selector.tsx
@@ -8,7 +8,8 @@ import { NumberInput } from '@/components/forms';
 import { GroupedAccountSelector } from '@/components/groupedaccountselector/groupedaccountselector';
 import { Amount } from '@/components/amount/amount';
 import { AmountUnit } from '@/components/amount/amount-with-unit';
-import { findAccount, getDisplayedCoinUnit } from '@/routes/account/utils';
+import { findAccount } from '@/routes/account/utils';
+import { getDisplayedCoinUnit } from '@/utils/coin';
 import { useMountedRef } from '@/hooks/mount';
 import style from './input-with-account-selector.module.css';
