+++ frontends/web/src/routes/market/swap/swap.tsx
@@ -38,7 +38,8 @@ import { DesktopBackButton } from '@/components/backbutton/backbutton';
 import { AmountWithUnit } from '@/components/amount/amount-with-unit';
 import { ArrowSwap } from '@/components/icon';
 import { SpinnerRingAnimated } from '@/components/spinner/SpinnerAnimation';
-import { findAccount, getDisplayedCoinUnit, isBitcoinOnly } from '@/routes/account/utils';
+import { findAccount } from '@/routes/account/utils';
+import { getDisplayedCoinUnit, isBitcoinOnly } from '@/utils/coin';
 import { useLoad } from '@/hooks/api';
 import { InputWithAccountSelector } from './components/input-with-account-selector';
 import { SwapServiceSelector } from './components/swap-service-selector';
